alter latencies table for column latency from float to decimal

+++ database/migrations/2024_04_04_202109_create_latencies_table.php
@@ -17,7 +17,7 @@ class CreateLatenciesTable extends Migration
             $table->increments('id');
             $table->integer('number_code')->nullable();
             $table->char('site')->require();
-            $table->float('latency')->nullable();
+            $table->decimal('latency', 16, 2)->nullable();
             $table->integer('status')->nullable();
             $table->timestamps();
         });
